+++ vec3.h
@@ -10,23 +10,23 @@ class Vec3
 {
 private:
 public:
-    float e[3];
+    double e[3];
 
     Vec3() : e{0, 0, 0} {}
 
-    Vec3(float e0, float e1, float e2) : e{e0, e1, e2} {}
+    Vec3(double e0, double e1, double e2) : e{e0, e1, e2} {}
 
-    float X() const { return e[0]; }
+    double X() const { return e[0]; }
 
-    float Y() const { return e[1]; }
+    double Y() const { return e[1]; }
 
-    float Z() const { return e[2]; }
+    double Z() const { return e[2]; }
 
     Vec3 operator-() const { return Vec3(-e[0], -e[1], -e[2]); }
 
-    float operator[](int i) const { return e[i]; }
+    double operator[](int i) const { return e[i]; }
 
-    float &operator[](int i) { return e[i]; }
+    double &operator[](int i) { return e[i]; }
 
     Vec3 &operator+=(const Vec3 &v)
     {
@@ -36,7 +36,7 @@ public:
         return *this;
     }
 
-    Vec3 &operator*=(float t)
+    Vec3 &operator*=(double t)
     {
         e[0] *= t;
         e[1] *= t;
@@ -44,17 +44,17 @@ public:
         return *this;
     }
 
-    Vec3 &operator/=(float t)
+    Vec3 &operator/=(double t)
     {
         return *this *= 1 / t;
     }
 
-    float Length() const
+    double Length() const
     {
         return sqrt(LengthSquared());
     }
 
-    float LengthSquared() const
+    double LengthSquared() const
     {
         return e[0] * e[0] + e[1] * e[1] + e[2] * e[2];
     }
@@ -84,22 +84,22 @@ inline Vec3 operator*(const Vec3 &u, const Vec3 &v)
     return Vec3(u.e[0] * v.e[0], u.e[1] * v.e[1], u.e[2] * v.e[2]);
 }
 
-inline Vec3 operator*(float t, const Vec3 &v)
+inline Vec3 operator*(double t, const Vec3 &v)
 {
     return Vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
-inline Vec3 operator*(const Vec3 &v, float t)
+inline Vec3 operator*(const Vec3 &v, double t)
 {
     return t * v;
 }
 
-inline Vec3 operator/(Vec3 v, float t)
+inline Vec3 operator/(Vec3 v, double t)
 {
     return (1 / t) * v;
 }
 
-inline float Dot(const Vec3 &u, const Vec3 &v)
+inline double Dot(const Vec3 &u, const Vec3 &v)
 {
     return u.e[0] * v.e[0] + u.e[1] * v.e[1] + u.e[2] * v.e[2];
 }
